Create script_setSharingNameAD.sh
Script for computer name remediation based on Active Directory Object. Script matches bound Active Directory object name to the current local Sharing computer name. If there is a match then the script will echo a result else if there is no match the script will sync the local Sharing computer name to the bound Active Directory object name and echo a result. This script also has the ability to log when the computer name sync occurs to a logfile of your choice.