Create ea_checkPointVer.sh
Extension Attribute to display Checkpoint VPN version. Scope as a Smart Group to all computers with Checkpoint VPN installed. This will display the current version name. The default script provided by Checkpoint uses cat, awk, sed, and is a mess. This script instead uses Plistbuddy which is much cleaner and gives the expected result with out cat, awk, and sed.